RenderObject::updateCull is now float, some Lua additions
- add obj_getUpdateCull()
- add obs parameter to getWallNormal(x, y, dist, obs) to control which obs to consider
- add getPerformanceCounter(), getPerformanceFreq()

+++ BBGE/RenderObject.h
@@ -267,7 +267,7 @@ public:
 	bool shareColorWithChildren;
 
 	bool cull;
-	int updateCull;
+	float updateCull;
 	int layer;
 
 	InterpolatedVector *positionSnapTo;
